# old_web_code/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for shift scheduler with alarms
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Shift {
  id: number;
  /** Shift name (e.g. Morning, Evening) */
  name: string;
  /** Hex color for display */
  color: string;
  /** Start time HH:MM */
  startTime: string;
  /** End time HH:MM */
  endTime: string;
  /** ISO date string YYYY-MM-DD */
  date: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateShiftBody {
  name: string;
  color: string;
  startTime: string;
  endTime: string;
  date: string;
  notes?: string | null;
}

export interface Alarm {
  id: number;
  /** Optional link to a shift */
  shiftId?: number | null;
  label: string;
  /** Time HH:MM */
  time: string;
  /** Days of week (0=Sun, 1=Mon ... 6=Sat) */
  daysOfWeek: number[];
  enabled: boolean;
  /** Minutes before shift start to fire alarm */
  minutesBefore?: number | null;
  createdAt: string;
}

export interface CreateAlarmBody {
  shiftId?: number | null;
  label: string;
  time: string;
  daysOfWeek: number[];
  enabled: boolean;
  minutesBefore?: number | null;
}

export interface UpdateAlarmBody {
  shiftId?: number | null;
  label?: string;
  time?: string;
  daysOfWeek?: number[];
  enabled?: boolean;
  minutesBefore?: number | null;
}

export interface ShiftSummary {
  totalShifts: number;
  shiftsThisWeek: number;
  shiftsThisMonth: number;
  totalHoursThisWeek: number;
  activeAlarms: number;
}

export type GetWeeklyShiftsParams = {
  /**
   * ISO date string for the start of the week (Monday)
   */
  weekStart: string;
};