# old_web_code/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for shift scheduler with alarms
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all shifts
 */
export const ListShiftsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string().describe("Shift name (e.g. Morning, Evening)"),
  color: zod.string().describe("Hex color for display"),
  startTime: zod.string().describe("Start time HH:MM"),
  endTime: zod.string().describe("End time HH:MM"),
  date: zod.string().describe("ISO date string YYYY-MM-DD"),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListShiftsResponse = zod.array(ListShiftsResponseItem);

/**
 * @summary Create a shift
 */
export const CreateShiftBody = zod.object({
  name: zod.string(),
  color: zod.string(),
  startTime: zod.string(),
  endTime: zod.string(),
  date: zod.string(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a shift by ID
 */
export const GetShiftParams = zod.object({
  id: zod.coerce.number(),
});

export const GetShiftResponse = zod.object({
  id: zod.number(),
  name: zod.string().describe("Shift name (e.g. Morning, Evening)"),
  color: zod.string().describe("Hex color for display"),
  startTime: zod.string().describe("Start time HH:MM"),
  endTime: zod.string().describe("End time HH:MM"),
  date: zod.string().describe("ISO date string YYYY-MM-DD"),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Update a shift
 */
export const UpdateShiftParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateShiftBody = zod.object({
  name: zod.string(),
  color: zod.string(),
  startTime: zod.string(),
  endTime: zod.string(),
  date: zod.string(),
  notes: zod.string().nullish(),
});

export const UpdateShiftResponse = zod.object({
  id: zod.number(),
  name: zod.string().describe("Shift name (e.g. Morning, Evening)"),
  color: zod.string().describe("Hex color for display"),
  startTime: zod.string().describe("Start time HH:MM"),
  endTime: zod.string().describe("End time HH:MM"),
  date: zod.string().describe("ISO date string YYYY-MM-DD"),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a shift
 */
export const DeleteShiftParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all alarms
 */
export const ListAlarmsResponseItem = zod.object({
  id: zod.number(),
  shiftId: zod.number().nullish().describe("Optional link to a shift"),
  label: zod.string(),
  time: zod.string().describe("Time HH:MM"),
  daysOfWeek: zod
    .array(zod.number())
    .describe("Days of week (0=Sun, 1=Mon ... 6=Sat)"),
  enabled: zod.boolean(),
  minutesBefore: zod
    .number()
    .nullish()
    .describe("Minutes before shift start to fire alarm"),
  createdAt: zod.string(),
});
export const ListAlarmsResponse = zod.array(ListAlarmsResponseItem);

/**
 * @summary Create an alarm
 */
export const CreateAlarmBody = zod.object({
  shiftId: zod.number().nullish(),
  label: zod.string(),
  time: zod.string(),
  daysOfWeek: zod.array(zod.number()),
  enabled: zod.boolean(),
  minutesBefore: zod.number().nullish(),
});

/**
 * @summary Update an alarm
 */
export const UpdateAlarmParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateAlarmBody = zod.object({
  shiftId: zod.number().nullish(),
  label: zod.string().optional(),
  time: zod.string().optional(),
  daysOfWeek: zod.array(zod.number()).optional(),
  enabled: zod.boolean().optional(),
  minutesBefore: zod.number().nullish(),
});

export const UpdateAlarmResponse = zod.object({
  id: zod.number(),
  shiftId: zod.number().nullish().describe("Optional link to a shift"),
  label: zod.string(),
  time: zod.string().describe("Time HH:MM"),
  daysOfWeek: zod
    .array(zod.number())
    .describe("Days of week (0=Sun, 1=Mon ... 6=Sat)"),
  enabled: zod.boolean(),
  minutesBefore: zod
    .number()
    .nullish()
    .describe("Minutes before shift start to fire alarm"),
  createdAt: zod.string(),
});

/**
 * @summary Delete an alarm
 */
export const DeleteAlarmParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get shifts for a specific week
 */
export const GetWeeklyShiftsQueryParams = zod.object({
  weekStart: zod.coerce.string(),
});

export const GetWeeklyShiftsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string().describe("Shift name (e.g. Morning, Evening)"),
  color: zod.string().describe("Hex color for display"),
  startTime: zod.string().describe("Start time HH:MM"),
  endTime: zod.string().describe("End time HH:MM"),
  date: zod.string().describe("ISO date string YYYY-MM-DD"),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
});
export const GetWeeklyShiftsResponse = zod.array(GetWeeklyShiftsResponseItem);

/**
 * @summary Get shift summary statistics
 */
export const GetShiftSummaryResponse = zod.object({
  totalShifts: zod.number(),
  shiftsThisWeek: zod.number(),
  shiftsThisMonth: zod.number(),
  totalHoursThisWeek: zod.number(),
  activeAlarms: zod.number(),
});