# old_web_code/lib/api-zod/src/generated/types/alarm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for shift scheduler with alarms
 * OpenAPI spec version: 0.1.0
 */

export interface Alarm {
  id: number;
  /** Optional link to a shift */
  shiftId?: number | null;
  label: string;
  /** Time HH:MM */
  time: string;
  /** Days of week (0=Sun, 1=Mon ... 6=Sat) */
  daysOfWeek: number[];
  enabled: boolean;
  /** Minutes before shift start to fire alarm */
  minutesBefore?: number | null;
  createdAt: string;
}