# old_web_code/lib/api-zod/src/generated/types/createShiftBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for shift scheduler with alarms
 * OpenAPI spec version: 0.1.0
 */

export interface CreateShiftBody {
  name: string;
  color: string;
  startTime: string;
  endTime: string;
  date: string;
  notes?: string | null;
}