# old_web_code/lib/api-zod/src/generated/types/shift.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for shift scheduler with alarms
 * OpenAPI spec version: 0.1.0
 */

export interface Shift {
  id: number;
  /** Shift name (e.g. Morning, Evening) */
  name: string;
  /** Hex color for display */
  color: string;
  /** Start time HH:MM */
  startTime: string;
  /** End time HH:MM */
  endTime: string;
  /** ISO date string YYYY-MM-DD */
  date: string;
  notes?: string | null;
  createdAt: string;
}