# old_web_code/lib/api-zod/src/generated/types/updateAlarmBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for shift scheduler with alarms
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAlarmBody {
  shiftId?: number | null;
  label?: string;
  time?: string;
  daysOfWeek?: number[];
  enabled?: boolean;
  minutesBefore?: number | null;
}